moved types to a separate file
